// web/src/api/generated/models/CollectionItemDetail.ts
export interface CollectionItemDetail {
    /**
     * 
     * @type {ResourceCapabilities}
     * @memberof CollectionItemDetail
     */
    capabilities: ResourceCapabilities;
    /**
     * 
     * @type {string}
     * @memberof CollectionItemDetail
     */
    collectionId: string;
    /**
     * 
     * @type {boolean}
     * @memberof CollectionItemDetail
     */
    completed: boolean;
    /**
     * 
     * @type {Date}
     * @memberof CollectionItemDetail
     */
    createdAt: Date;
    /**
     * 
     * @type {string}
     * @memberof CollectionItemDetail
     */
    createdBy: string;
    /**
     * 
     * @type {AuthorSummary}
     * @memberof CollectionItemDetail
     */
    creator: AuthorSummary;
    /**
     * 
     * @type {string}
     * @memberof CollectionItemDetail
     */
    id: string;
    /**
     * 
     * @type {number}
     * @memberof CollectionItemDetail
     */
    position: number;
    /**
     * 
     * @type {string}
     * @memberof CollectionItemDetail
     */
    title: string;
    /**
     * 
     * @type {Date}
     * @memberof CollectionItemDetail
     */
    updatedAt: Date;
    /**
     * 
     * @type {number}
     * @memberof CollectionItemDetail
     */
    version: number;
}